On buy request return remain amount, send machine address in eris to coins server

+++ src/main/java/com/softjourn/vending/entity/Purchase.java
@@ -16,7 +16,7 @@ public class Purchase {
 
     @Id
     @GeneratedValue
-    private Double id;
+    private Long id;
 
     @Column
     private String account;
